requirements.txt
Django==3.0.5
python-dotenv==0.12.0
requests==2.23.0
